Implement Doctrine ORM.
This commit marks the beginning of a heavily refactoring of our amp-framework.
We will now use Doctrine ORM for database-handling stuff and additionally
preparations have been done to publish ampf on packagist. More heavily changes will
follow, so this is just the start. This commit marks the beginning of the 2.0-branch.